Refactored Models to follow the updated diagrams. On that note, the existing relationship between User and Student, Employer and Staff was changed into a joined-table inheritance. The relationships between models was also refactored to more strictly adhere to diagram and business logic of the application as understood so for
Implemented the state design pattern, however, did not implement the functionality of each state because proper testing of this requries the controller functions to be implemented

NOTE: Fixed the visible syntax errors that occured as a result of this refactor. Althought the application will run "correctly" and initialize upon calling flask init, there may be other issues such as logic errors that the IDE will not immediately flag on runtime.